requirements.txt
discord.py[voice]>=1.7.2
gTTS==2.2.3